Add ability to edit members and groups

+++ app/src/main/java/org/groover/bar/util/app/ItemList.kt
@@ -76,7 +76,7 @@ fun ItemListEntry(item: Item, index: Int, amount: Int, setAmount: (Int) -> Unit)
             ) { Text("-") }
 
             Text(
-                amount.toInt().toString(),
+                amount.toString(),
                 modifier = Modifier.padding(horizontal = 15.dp)
             )
